+++ core/src/main/kotlin/net/corda/core/node/NodeInfo.kt
@@ -15,7 +15,7 @@ import java.security.cert.X509Certificate
  * The identity can be used in flows and is distinct from the Node's legalIdentity
  */
 @CordaSerializable
-data class ServiceEntry(val info: ServiceInfo, val identity: Party)
+data class ServiceEntry(val info: ServiceInfo, val identity: PartyAndCertificate)
 
 /**
  * Info about a network node that acts on behalf of some form of contract party.
@@ -30,6 +30,6 @@ data class NodeInfo(val address: SingleMessageRecipient,
         require(advertisedServices.none { it.identity == legalIdentity }) { "Service identities must be different from node legal identity" }
     }
 
-    val notaryIdentity: Party get() = advertisedServices.single { it.info.type.isNotary() }.identity
+    val notaryIdentity: PartyAndCertificate get() = advertisedServices.single { it.info.type.isNotary() }.identity
     fun serviceIdentities(type: ServiceType): List<Party> = advertisedServices.filter { it.info.type.isSubTypeOf(type) }.map { it.identity }
 }